requirements.txt
requests
beautifulsoup4
attr
aiohttp

# emit
python-twitter
twilio

# testing
pytest
aresponses
responses
mypy
pytest-mypy

sqlalchemy
